+++ packages/error/src/UnimplementedError.ts
@@ -1,4 +1,4 @@
-import { Ambiguous } from '@jamashita/anden-type/src/index';
+import { Ambiguous } from '@jamashita/anden-type';
 import { RuntimeError } from './RuntimeError';
 
 export class UnimplementedError extends RuntimeError {