+++ packages/error/src/mock/MockRuntimeError.ts
@@ -1,7 +1,8 @@
+import { Ambiguous } from '@jamashita/anden-type';
 import { RuntimeError } from '../RuntimeError';
 
 export class MockRuntimeError extends RuntimeError {
-  public constructor(cause?: Error) {
+  public constructor(cause?: Ambiguous<Error>) {
     super('failed', cause);
   }
 }