+++ packages/ulid/src/ULIDError.ts
@@ -1,7 +1,8 @@
 import { RuntimeError } from '@jamashita/anden-error';
+import { Ambiguous } from '@jamashita/anden-type';
 
 export class ULIDError extends RuntimeError {
-  public constructor(message: string, cause?: Error) {
+  public constructor(message: string, cause?: Ambiguous<Error>) {
     super(message, cause);
   }
 }